requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0

# Traitement d'images
Pillow==10.1.0
numpy==1.24.3

# Utilitaires
pydantic==2.5.0
python-dotenv==1.0.0

# CORS et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4